Implement `Relation.cast` (#37)

+++ tests/test_duckdb/test_database.py
@@ -111,8 +111,8 @@ def test_database_create_table():
         "enum_column",
     ]
     assert list(table.types.values()) == [
-        "BIGINT",
-        "BIGINT",
+        "INTEGER",
+        "INTEGER",
         "VARCHAR",
         "VARCHAR",
         "BOOLEAN",
